fix and speed up trim null bytes code
The previous code will runtime panic if we
happen to have a byte slice with only 0x00.
And it also resizes the byte slice in every
step hence quite slow.

Speed it up and make it correct by extracting
it out to a helper function. It shaves off 2 seconds
from our testing code.

Fixes #88