fix and test distracting invalid null bytes trim
Fixes #88

The previous trimming code was
```go
for bz[0] == 0x00 {
    bz = bz[1:]
}
```
after being generated by a sha256 sum.

This code was flagged independently by both @veorq
and myself. @veorq analyzed and found that there
are about: 2^54 strings that'll trigger the panic
i.e. contain all zeros and cause a runtime panic.
If we process billions of names, the probability
is negligible.

Nonetheless this is code we shouldn't have in
the repo as it could get copied naively, it is slow too
plus it captured the attention of two independent auditors.

Fix this by using the stdlib's bytes.TrimLeft and also
include tests to lock this behavior in.