Only include prefix/type-wrapper in Marshal* or for interface values (not otherwise) (#160)
This PR fixes a major bug in the way Amino:binary was handling the encoding of concrete registered types as struct field values, when the type of that field is not an Interface. This was due to a bug in the structure of the code, carried over from the previous design of Amino (before it became more like Proto3).

I've come to see that as an anti-pattern. It is useful because it allows for the unmarshalling of a concrete type into an interface variable (since the prefix bytes serve to identify the concrete type to deserialize to). But it isn't good when you want to register common types like primitive types that don't need to be prefixed by the prefix bytes in most situations.

This PR allows us to preserve the functionality for marshalling a concrete type and unmarshalling it to an interface -- those top-level Marshal* and Unmarshal* functions will include prefix bytes (or in the case of Amino:JSON the type wrapper bytes) for all registered concrete types. But where the registered concrete type appears say as a struct field whose type is exactly that concrete type, the prefix bytes (nor type wrapper) are not included.

It's obvious in hindsight...