WIP proto3 compatibility (#163)
This PR makes Amino:binary compatible with Proto3 (but not Proto2). It doesn't implement all of the features of Proto3, but everything should be parseable in Proto3.

Notable differences are, support of Interfaces as byte-length prefixed blobs. Support for nested arrays is also interesting.

This makes the encoder slower, but the purpose is to start with Proto3 compatibility and work toward Amin2.

Many thanks to @ebuchman and @mdyring for exploring this.